Improved flexbox performance
Also added optional useWebWorker setting for the flexbox plugin. This will load the yoga library in a web worker and do the
calculations on the worker. At this moment, the postMessage command is too slow and will require some more work, but
you may get some performance gains from offloading the calculations.